Build the MIDI driver as both 32 and 64 bit.
It's possible an old 32-bit-only driver forces the MIDI Server
into 32-bit mode, even on post-10.7 OS X.

+++ Frameworks/SnoizeMIDISpy/Driver/MessagePortBroadcaster.cpp
@@ -255,7 +255,7 @@ void	MessagePortBroadcaster::AddListener(CFDataRef listenerIdentifierData)
     if (!listenerIdentifierNumber)
         return;
 
-    listenerPortName = CFStringCreateWithFormat(kCFAllocatorDefault, NULL, CFSTR("%@-%d"), mBroadcasterName, listenerIdentifier);
+    listenerPortName = CFStringCreateWithFormat(kCFAllocatorDefault, NULL, CFSTR("%@-%d"), mBroadcasterName, (int)listenerIdentifier);
 
     remotePort = CFMessagePortCreateRemote(kCFAllocatorDefault, listenerPortName);
     if (remotePort) {
